kleaf: Add support for building akita with kleaf
This creates the necessary bazel build files to compile the akita build
target with kleaf.

One unique difference is we are using label-typed build settings to
select the kernel build and SOC module for the build target. These
labels are defined in //private/google-modules/soc/gs as:

  * gs_kernel_build
  * gs_soc_module

and allow one to define the kernel_build and SoC module dependencies
within external modules to a generic label-typed build setting. For
akita builds, use --config=akita to set these labels to the zuma_akita
and zuma_soc.akita starlarks accordingly. This is defined in the
device.bazelrc file.

Change-Id: Ia90a66732c3c4062fe5199bf6cac345c8878cdb8
Signed-off-by: Will McVicker <willmcvicker@google.com>