ABI: add the update_symbol_list.sh script
This script can be used to update the pixel symbol list. Refer to the
usage message for details on the different command line arguments it
supports. The main value this script provides is cherry-picking the ABI
symbol list update from the aosp-staging/ kernel to the aosp/ kernel and
then rebasing that patch to the ToT so that developers can directly push
the symbol list patch to AOSP gerrit. The one caveat is that it doesn't
support cherry-picking any core kernel patches from aosp-staging/ to
aosp/. So if you need to make additional core kernel changes with the
symbol list patch, then apply those patches to the aosp/ kernel first
before running this script.

To update the symbol list, run this command from the root of the kernel
repo:

  ./update_symbol_list.sh --config DEVICE

To create a commit, use the --commit BUG_NUMBER argument. If you are
using the aosp-staging kernel, then this argument will also cherry-pick
the symbol list commit to the ToT AOSP kernel under the aosp/ folder.
For example:

  ./update_symbol_list.sh --config DEVICE --commit BUG_NUMBER

Change-Id: I9ce565bca3154642bcb8fabedcf92542cf440c57
Signed-off-by: Will McVicker <willmcvicker@google.com>