Update update_symbol_list.sh to only apply new symbol additions
Since aosp-staging includes symbols that don't exist in the public AOSP
kernel, we can't merge the aosp-staging/ and aosp/ pixel symbol lists
since that will add the staging-only symbols to the public aosp kernel
which isn't allowed. To fix this, only apply the symbol additions made.

For staging-only changes, don't use the --commit flag so that the change
isn't applied to the AOSP branch.

Bug: 274944767
Change-Id: I4702d0fe267e4b6ea04aa2c3d0b021348a78cb62
Signed-off-by: Will McVicker <willmcvicker@google.com>