update_symbol_list.sh: remove codenames
We technically don't need to list the codename in this script. We can
just check if the directoy private/devices/google/${TARGET} exists to
verify a valid target has been passed in. This allows us to move the
script to the private/google-modules/soc/gs git project so that it can
be used for all repo init partner groups.

Bug: 274961291
Change-Id: If2f8a33e917ed7579b9b425e4641143ec48f9b99
Signed-off-by: Will McVicker <willmcvicker@google.com>